+++ src/brod_group_subscriber.erl
@@ -79,7 +79,7 @@
 %%
 %% While this callback function is being evaluated, the fetch-ahead
 %% partition-consumers are fetching more messages behind the scene
-%% unless prefetch_count is set to 0 in consumer config.
+%% unless prefetch_count and prefetch_bytes are set to 0 in consumer config.
 %%
 -callback handle_message(brod:topic(),
                          brod:partition(),